+++ planning/autoware_mission_planner/src/mission_planner/mission_planner.hpp
@@ -19,18 +19,10 @@
 #include "arrival_checker.hpp"
 #include "reroute_safety.hpp"
 
-#include <autoware/component_interface_specs/planning.hpp>
-#include <autoware/component_interface_utils/rclcpp.hpp>
-#include <autoware_utils_logging/logger_level_configure.hpp>
-#include <autoware_utils_system/stop_watch.hpp>
 #include <autoware_vehicle_info_utils/vehicle_info_utils.hpp>
-#include <rclcpp/rclcpp.hpp>
-#include <tf2_ros/buffer.hpp>
-#include <tf2_ros/transform_listener.hpp>
 
 #include <autoware_adapi_v1_msgs/msg/operation_mode_state.hpp>
 #include <autoware_common_msgs/msg/response_status.hpp>
-#include <autoware_internal_debug_msgs/msg/float64_stamped.hpp>
 #include <autoware_map_msgs/msg/lanelet_map_bin.hpp>
 #include <autoware_planning_msgs/msg/lanelet_route.hpp>
 #include <autoware_planning_msgs/msg/route_state.hpp>
@@ -39,7 +31,6 @@
 #include <autoware_planning_msgs/srv/set_waypoint_route.hpp>
 #include <geometry_msgs/msg/transform_stamped.hpp>
 #include <nav_msgs/msg/odometry.hpp>
-#include <visualization_msgs/msg/marker_array.hpp>
 
 #include <functional>
 #include <memory>
@@ -59,12 +50,6 @@ using autoware_planning_msgs::srv::SetWaypointRoute;
 using geometry_msgs::msg::Pose;
 using nav_msgs::msg::Odometry;
 using visualization_msgs::msg::MarkerArray;
-
-using RouteStateSpecs = autoware::component_interface_specs::planning::RouteState;
-using ClearRouteSpecs = autoware::component_interface_specs::planning::ClearRoute;
-using SetLaneletRouteSpecs = autoware::component_interface_specs::planning::SetLaneletRoute;
-using SetWaypointRouteSpecs = autoware::component_interface_specs::planning::SetWaypointRoute;
-using LaneletRouteSpecs = autoware::component_interface_specs::planning::LaneletRoute;
 
 struct MissionPlannerConfig
 {
@@ -161,72 +146,6 @@ private:
   bool allow_reroute_in_autonomous_mode_;
   RerouteSafetyResult check_reroute_safety(
     const LaneletRoute & original_route, const LaneletRoute & target_route);
-};
-
-class MissionPlannerNode : public rclcpp::Node
-{
-public:
-  explicit MissionPlannerNode(const rclcpp::NodeOptions & options);
-
-private:
-  // Publishes the processing time on destruction, regardless of which return path is taken.
-  class ScopedProcessingTimePublisher
-  {
-  public:
-    explicit ScopedProcessingTimePublisher(MissionPlannerNode & node) : node_(node) {}
-    ~ScopedProcessingTimePublisher() { node_.publish_processing_time(stop_watch_); }
-
-  private:
-    MissionPlannerNode & node_;
-    autoware_utils_system::StopWatch<std::chrono::milliseconds> stop_watch_;
-  };
-
-  void publish_processing_time(
-    autoware_utils_system::StopWatch<std::chrono::milliseconds> stop_watch);
-
-  MissionPlanner mission_planner_;
-
-  std::string map_frame_;
-  tf2_ros::Buffer tf_buffer_;
-  tf2_ros::TransformListener tf_listener_;
-
-  autoware::component_interface_utils::NodeAdaptor<rclcpp::Node> adaptor_{this};
-  autoware::component_interface_utils::Service<ClearRouteSpecs>::SharedPtr srv_clear_route;
-  autoware::component_interface_utils::Service<SetLaneletRouteSpecs>::SharedPtr
-    srv_set_lanelet_route;
-  autoware::component_interface_utils::Service<SetWaypointRouteSpecs>::SharedPtr
-    srv_set_waypoint_route;
-  autoware::component_interface_utils::Publisher<RouteStateSpecs>::SharedPtr pub_state_;
-  autoware::component_interface_utils::Publisher<LaneletRouteSpecs>::SharedPtr pub_route_;
-
-  rclcpp::Subscription<PoseWithUuidStamped>::SharedPtr sub_modified_goal_;
-  rclcpp::Subscription<Odometry>::SharedPtr sub_odometry_;
-  rclcpp::Subscription<OperationModeState>::SharedPtr sub_operation_mode_state_;
-
-  rclcpp::Subscription<LaneletMapBin>::SharedPtr sub_vector_map_;
-  rclcpp::Publisher<MarkerArray>::SharedPtr pub_marker_;
-  rclcpp::Publisher<MarkerArray>::SharedPtr pub_goal_footprint_marker_;
-
-  void on_odometry(const Odometry::ConstSharedPtr msg);
-  void on_operation_mode_state(const OperationModeState::ConstSharedPtr msg);
-  void on_map(const LaneletMapBin::ConstSharedPtr msg);
-
-  void on_clear_route(
-    const ClearRoute::Request::SharedPtr req, const ClearRoute::Response::SharedPtr res);
-  void on_set_lanelet_route(
-    const SetLaneletRoute::Request::SharedPtr req, const SetLaneletRoute::Response::SharedPtr res);
-  void on_set_waypoint_route(
-    const SetWaypointRoute::Request::SharedPtr req,
-    const SetWaypointRoute::Response::SharedPtr res);
-
-  void publish_pose_log(const Pose & pose, const std::string & pose_type);
-
-  rclcpp::TimerBase::SharedPtr data_check_timer_;
-  void check_initialization();
-
-  std::unique_ptr<autoware_utils_logging::LoggerLevelConfigure> logger_configure_;
-  rclcpp::Publisher<autoware_internal_debug_msgs::msg::Float64Stamped>::SharedPtr
-    pub_processing_time_;
 };
 
 }  // namespace autoware::mission_planner
