+++ planning/autoware_mission_planner/test/test_mission_planner_node.cpp
@@ -12,7 +12,7 @@
 // See the License for the specific language governing permissions and
 // limitations under the License.
 
-#include "../src/mission_planner/mission_planner.hpp"
+#include "../src/mission_planner/mission_planner_node.hpp"
 
 #include <ament_index_cpp/get_package_share_directory.hpp>
 #include <autoware/lanelet2_utils/conversion.hpp>
